Generators config and class

+++ src/main/java/pl/ynfuien/ygenerators/core/Doubledrop.java
@@ -39,63 +39,46 @@ public class Doubledrop {
         return database.setDoubledrop(timeLeft, multiplayer);
     }
 
-    // Gets whether double drop is active
     public boolean isActive() {
         return timeLeft == -1 || timeLeft > 0;
     }
 
-    // Gets double drop time left
     public long getTimeLeft() {
         return timeLeft;
     }
 
-    // Gets formatted double drop time left
     @NotNull
     public String getFormattedTimeLeft() {
-        // Return infinity time format if double drop time is infinity
         if (timeLeft == -1) return Lang.Message.DOUBLEDROP_TIME_INFINITY.get();
 
-        // Get hours from time left
-        long hours = timeLeft / 60;
-        // Get minutes from time left
+        int hours = timeLeft / 60;
         int minutes = timeLeft % 60;
 
-        // Create hashmap for placeholders
         HashMap<String, Object> placeholders = new HashMap<>();
-        // Add hours to placeholders
         placeholders.put("hours", hours);
-        // Add minutes to placeholders
         placeholders.put("minutes", minutes);
 
-        // Return time
         return Lang.Message.DOUBLEDROP_TIME.get(placeholders);
     }
 
-    // Gets double drop multiplayer
     public double getMultiplayer() {
         return multiplayer;
     }
 
-    // Adds time to double drop time left
     public void addTime(int time) {
         if (timeLeft == -1) return;
         setTimeLeft(timeLeft + time);
     }
 
-    // Sets double drop time left
     public void setTimeLeft(int timeLeft) {
-        // Set time left to -1 if it is lower than -1
         if (timeLeft < -1) timeLeft = -1;
         this.timeLeft = timeLeft;
 
-        // Save new time in config
         save();
-        // Return if time left is lower than 1
         if (timeLeft < 1) {
             cancelInterval();
             return;
         }
-        // Return if interval is already set
         if (intervalTask != null) return;
 
         intervalTask = Bukkit.getAsyncScheduler().runAtFixedRate(instance, (task) -> {
@@ -129,14 +112,12 @@ public class Doubledrop {
         return true;
     }
 
-    // Removes time from double drop time left
     public void removeTime(int time) {
         if (timeLeft == -1) return;
         if (timeLeft - time < 0) time = timeLeft;
         setTimeLeft(timeLeft - time);
     }
 
-    // Sets double drop multiplayer
     public void setMultiplayer(float multiplayer) {
         if (multiplayer < 0) multiplayer = 0;
         this.multiplayer = multiplayer;
